Eliminate sys absent refl before applying cb_op_to_min
If reflections are present that should be systematically absent according to
centring this will lead to a "Change of basis yields non-integral Miller index"
error.

Fixes #1064.
