+++ newsfragments/1064.bugfix
@@ -1,0 +1,2 @@
+Eliminate systematic absences before applying change of basis op to minimum 
+cell in dials.symmetry.
